added more analytics
added analytics for player movement and for each upgrade to track what upgrades are being used
also added a start complete and fail progression status to track when the player starts when they quit and when they die.

also added in litter objects on the floor for more detail

also updated the inventory, tab now opens the inventory and clicking on the icons now uses/switches between the two guninventory and the inventory.

added ui elements to show how to open the shop and inventory.

updated ui and movement Script